feat: FeaturePersistencePort 추가 (CM-105)
