+++ src/main/java/com/starmix/checkmate/adapter/in/sse/project/response/CreateFeatureSpecificationResponse.java
@@ -1,6 +1,6 @@
 package com.starmix.checkmate.adapter.in.sse.project.response;
 
-import com.starmix.checkmate.domain.project.Feature;
+import com.starmix.checkmate.domain.feature.Feature;
 import lombok.Builder;
 
 import java.util.List;
