+++ src/main/java/com/starmix/checkmate/adapter/in/sse/project/response/FeedbackResponse.java
@@ -1,7 +1,7 @@
 package com.starmix.checkmate.adapter.in.sse.project.response;
 
 import com.starmix.checkmate.adapter.out.ai.dto.FeedbackDto;
-import com.starmix.checkmate.domain.project.Feature;
+import com.starmix.checkmate.domain.feature.Feature;
 import lombok.Builder;
 
 import java.util.List;
