+++ src/main/java/com/starmix/checkmate/domain/feature/Feature.java
@@ -1,0 +1,31 @@
+package com.starmix.checkmate.domain.feature;
+
+import com.starmix.checkmate.domain.common.Stack;
+import com.starmix.checkmate.domain.task.Priority;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@Builder(toBuilder = true)
+public class Feature {
+    private final String featureId;
+    private final String name;
+    private final String useCase;
+    private final String input;
+    private final String output;
+    private final String preCondition;
+    private final String postCondition;
+    private final List<Stack> stacks;
+    private final Time time;
+    private final Integer difficulty;
+    private final Priority priority;
+    private final String projectId;
+
+    public static Feature fromFeatureName(String featureName) {
+        return Feature.builder()
+                .name(featureName)
+                .build();
+    }
+}
