+++ src/main/java/com/starmix/checkmate/domain/feature/Time.java
@@ -1,0 +1,10 @@
+package com.starmix.checkmate.domain.feature;
+
+import java.time.LocalDate;
+
+public record Time(
+        Integer expectedDays,
+        LocalDate startDate,
+        LocalDate endDate
+) {
+}
