--- src/main/java/com/starmix/checkmate/domain/project/Feature.java
@@ -1,19 +1,0 @@
-package com.starmix.checkmate.domain.project;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder(toBuilder = true)
-public class Feature {
-    private final String name;
-    private final String useCase;
-    private final String input;
-    private final String output;
-
-    public static Feature fromFeatureName(String featureName) {
-        return Feature.builder()
-                .name(featureName)
-                .build();
-    }
-}
